Conflict Resolved FIle Chaged categoryController ,ICategoriesRepository,CategoryRepository,startup.cs
