+++ Trappist/src/Promact.Trappist.Repository/Categories/CategoryRepository.cs
@@ -1,4 +1,6 @@
 ﻿using Promact.Trappist.DomainModel.DbContext;
+using Promact.Trappist.DomainModel.Models.Category;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -21,5 +23,49 @@ namespace Promact.Trappist.Repository.Categories
             var category = _dbContext.Category.Select(x => x.CategoryName).ToList();
             return (category);
         }
+
+        #region Adding a CategoryName
+        /// <summary>
+        /// Adding a Category in Category model
+        /// </summary>
+        /// <param name="catagory">Object of class Category</param>
+        public void AddCategory(Category category)
+        {
+            _dbContext.Category.Add(category);
+            _dbContext.SaveChanges();
+        }
+        #endregion
+
+        #region Finding a Id Respective Category
+        /// <summary>
+        /// Find a Respective Id from Catagory Table
+        /// </summary>
+        /// <param name="Key"></param>
+        /// <Returns>if key foundthen Return respective category from category table or will return Null</Returns>
+        public Category Getcategory(int key)
+        {
+            try
+            {
+                return _dbContext.Category.FirstOrDefault(Check => Check.Id == key);
+            }
+            catch (Exception)
+            {
+                return null;
+            }
+        }
+        #endregion
+
+        #region Edit A Category Name
+        // <summary>
+        // Edit a Category from Category Table
+        // </summary>
+        // <param name="catagory">object of the class Category</param>
+        public void CategoryEdit(Category category)
+        {
+            _dbContext.Category.Update(category);
+            _dbContext.SaveChanges();
+        }
+        #endregion
     }
 }
+
