+++ Trappist/src/Promact.Trappist.Repository/Categories/ICategoryRepository.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using Promact.Trappist.DomainModel.Models.Category;
+using System.Collections.Generic;
 
 namespace Promact.Trappist.Repository.Categories
 {
@@ -9,5 +10,22 @@ namespace Promact.Trappist.Repository.Categories
         /// </summary>
         /// <returns>Category list</returns>
         IEnumerable<string> GetAllCategories();
+
+        /// <summary>
+        /// Add a category in Category Model
+        /// </summary>
+        /// <param name="catagory"></param>
+        void AddCategory(Category catagory);
+        /// <summary>
+        /// it will search Id and corresponding category
+        /// </summary>
+        /// <param name="key">unique key of a table</param>
+        /// <returns>if Id match then it will return object</returns>
+        Category Getcategory(int key);
+        /// <summary>
+        /// Edit category From Category model
+        /// </summary>
+        /// <param name="catagory"> Object of class Category</param>
+        void CategoryEdit(Category catagory);
     }
 }
